[MISC] Remove extra header previous_level.hpp
