--- include/chopper/layout/previous_level.hpp
@@ -1,30 +1,0 @@
-#pragma once
-
-#include <cassert>
-#include <string>
-#include <vector>
-
-namespace chopper::layout
-{
-
-//!\brief Information about the previous IBF level to be passed down to ensure correct output.
-struct previous_level
-{
-    std::vector<size_t> bin_indices{};
-    std::string bin_indices_str;
-    std::string num_of_bins;
-    std::string estimated_sizes;
-    std::string optimal_score;
-    std::string correction;
-    std::string tmax;
-
-    bool empty() const
-    {
-        assert(((((bin_indices.empty() == num_of_bins.empty()) == estimated_sizes.empty()) == optimal_score.empty())
-                == correction.empty())
-               == tmax.empty());
-        return bin_indices.empty();
-    }
-};
-
-} // namespace chopper::layout
